requirements.txt
python-telegram-bot[job-queue]==20.7
requests==2.31.0
python-dotenv==1.0.0
pytz==2024.1
Flask==3.0.0
pytz==2024.1